requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
pydantic==2.12.3
pydantic-settings==2.11.0
python-multipart==0.0.20
python-dotenv==1.1.1


openai==1.83.0

# Additional dependencies
aiohttp==3.11.11
nest-asyncio==1.6.0

# Google Gemini
google-genai
